Created the MVVM folders and displaying dates on screen
