--- AsyncAwaitExample/ContentView.swift
@@ -1,74 +1,0 @@
-//
-//  ContentView.swift
-//  AsyncAwaitExample
-//
-//  Created by EstrHuP on 13/6/23.
-//
-
-import SwiftUI
-
-struct CurrentDate: Decodable, Identifiable {
-    let id = UUID()
-    let date: String
-    
-    private enum CodingKeys: String, CodingKey {
-        case date = "date"
-    }
-}
-
-struct ContentView: View {
-    
-    @State private var currentDates: [CurrentDate] = []
-
-    private func getDate() async throws -> CurrentDate? {
-        
-        guard let url = URL(string: "https://ember-sparkly-rule.glitch.me/current-date") else {
-            fatalError("Url is incorrect!")
-        }
-        
-        let (data, _) = try await URLSession.shared.data(from: url)
-        return try? JSONDecoder().decode(CurrentDate.self, from: data)
-    }
-    
-    private func populateDates() async {
-        //Implements async await
-        do {
-            guard let currentDate = try await getDate() else {
-                return
-            }
-            //Save the current date when getDate finished
-            self.currentDates.append(currentDate)
-        } catch {
-            print("error")
-        }
-    }
-    
-    var body: some View {
-        NavigationView {
-            List(currentDates) { currentDate in
-                Text(currentDate.date)
-            }.listStyle(.plain)
-            
-            .navigationTitle("Dates")
-            .navigationBarItems(trailing: Button(action: {
-                // async { await populateDates() }
-                //iOS 15
-                Task.init(operation: {
-                    await populateDates()
-                })
-            }, label: {
-                Image(systemName: "arrow.clockwise.circle")
-            }))
-        }
-        //iOS 15
-        .task {
-            await populateDates()
-        }
-    }
-}
-
-struct ContentView_Previews: PreviewProvider {
-    static var previews: some View {
-        ContentView()
-    }
-}
