+++ AsyncAwaitExample/Example-CurrentDate/WebService.swift
@@ -9,7 +9,7 @@ import Foundation
 
 class WebService {
     
-    private func getDate() async throws -> CurrentDate? {
+    func getDate() async throws -> CurrentDate? {
         
         guard let url = URL(string: "https://ember-sparkly-rule.glitch.me/current-date") else {
             fatalError("Url is incorrect!")
